+++ R/occumbData.R
@@ -85,6 +85,19 @@ validate_occumbData <- function(object) {
                          knitr::combine_words(names(object@repl_cov)[wrong_repl_cov],
                                               before = "'", after = "'")))
 
+    ## Appropriate covariate classes.
+    valid_class <- c("logical", "numeric", "integer", "factor", "character", "NULL")
+    cov_class <- c(sapply(object@spec_cov, class),
+                   sapply(object@site_cov, class),
+                   sapply(object@repl_cov, function(x) class(c(x))))
+    if (any(cov_class %!in% valid_class))
+        msg <- c(
+            msg,
+            sprintf("%s must be logical, numeric, integer, factor, or character",
+                    knitr::combine_words(names(cov_class[cov_class %!in% valid_class]),
+                                         before = "'", after = "'"))
+        )
+
     ## No missing values in covariates.
     if (!checkmate::test_vector(unlist(object@spec_cov),
                                 any.missing = FALSE,
@@ -191,55 +204,11 @@ occumbData <- function(y,
                        site_cov = NULL,
                        repl_cov = NULL) {
 
-    # The mode of covariates is numeric, factor, or character.
-    check_covariate_mode(spec_cov, site_cov, repl_cov)
-
     out <- methods::new("occumbData",
                         y = y,
                         spec_cov = spec_cov,
                         site_cov = site_cov,
                         repl_cov = repl_cov)
     return(out)
-}
-
-# Check for the mode of each covariate
-check_covariate_mode <- function(spec_cov, site_cov, repl_cov) {
-    wrong_spec_cov_mode <- vector(length = length(spec_cov))
-    for (i in seq_along(spec_cov)) {
-        if (!(mode(spec_cov[[i]]) == "numeric" |
-              mode(spec_cov[[i]]) == "factor" | 
-              mode(spec_cov[[i]]) == "character"))
-            wrong_spec_cov_mode[i] <- TRUE
-    }
-    wrong_site_cov_mode <- vector(length = length(site_cov))
-    for (i in seq_along(site_cov)) {
-        if (!(mode(site_cov[[i]]) == "numeric" |
-              mode(site_cov[[i]]) == "factor" | 
-              mode(site_cov[[i]]) == "character"))
-            wrong_site_cov_mode[i] <- TRUE
-    }
-    wrong_repl_cov_mode <- vector(length = length(repl_cov))
-    for (i in seq_along(repl_cov)) {
-        if (!(mode(repl_cov[[i]]) == "numeric" |
-              mode(repl_cov[[i]]) == "factor" | 
-              mode(repl_cov[[i]]) == "character"))
-            wrong_repl_cov_mode[i] <- TRUE
-    }
-
-    if (sum(c(wrong_spec_cov_mode, wrong_site_cov_mode, wrong_repl_cov_mode))) {
-        var <- c(names(spec_cov)[wrong_spec_cov_mode],
-                 names(site_cov)[wrong_site_cov_mode],
-                 names(repl_cov)[wrong_repl_cov_mode])
-        mod <- NULL
-        for (i in seq_len(sum(wrong_spec_cov_mode)))
-            mod <- c(mod, mode(spec_cov[[which(wrong_spec_cov_mode)[i]]]))
-        for (i in seq_len(sum(wrong_site_cov_mode)))
-            mod <- c(mod, mode(site_cov[[which(wrong_site_cov_mode)[i]]]))
-        for (i in seq_len(sum(wrong_repl_cov_mode)))
-            mod <- c(mod, mode(repl_cov[[which(wrong_repl_cov_mode)[i]]]))
-
-        stop(message = sprintf("Unacceptable mode: the following covariates must be numeric, factor, or character. \n %s",
-                               paste(sprintf("%s: %s", var, mod), collapse = "; ")))
-    }
 }